break out some components and rename/organize others WIP
