+++ src/App.js
@@ -1,16 +1,21 @@
 import React from "react";
 import { BrowserRouter as Router, Route, Redirect } from "react-router-dom";
 import "./App.css";
-import Tutorial from "./Tutorial";
-import Documentation from "./Documentation";
-import Header from "./Header";
-import Menu from "./Menu";
+import Tutorial from "./pages/Tutorial";
+import Documentation from "./pages/Documentation";
+import Title from "./title/Title";
+import Menu from "./drawer/Menu";
+import { Tabs, Divider } from "./components";
 
 const App = () => (
   <div className="App">
     <Menu />
     <Router>
-      <Header />
+      <header className="container">
+        <Title />
+        <Tabs />
+        <Divider />
+      </header>
       <main>
         <Route exact path="/" render={() => <Redirect to="/docs" />} />
         <Route path="/docs" component={Documentation} />
